packages/analysis/requirements.txt
fastapi==0.95.2
uvicorn>=0.20.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
mysql-connector-python>=8.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyecharts>=2.0.0
openpyxl>=3.1.0
python-multipart>=0.0.6
typing-extensions>=4.8.0
starlette==0.27.0
lifelines>=0.27.0
pycountry>=22.3.0
